Changed all uses of deltaTime to use float for time in seconds instead of int with milliseconds.
The Java implementation used milliseconds, but using the time as a cast to milliseconds causes disruption to normal flow when the numbers are too small.

+++ Uno-CPP/Uno-CPP/AIPlayer.cpp
@@ -13,7 +13,7 @@ AIPlayer::AIPlayer(const int playerID, const std::string & playerName, const sf:
 	_consideringDelayTimer = -1;
 }
 
-void AIPlayer::update(const int deltaTime)
+void AIPlayer::update(const float deltaTime)
 {
 	updateAntiUnoCheck(deltaTime);
 	updateJumpInCheck(deltaTime);
